requirements.txt
pandas>=1.2.2
google-cloud-storage>=1.36.0
google-cloud-logging==3.5.0